refactor(st-09052): modularize relational query builder
